+++ src/components/index.js
@@ -1,5 +1,5 @@
 export { default as Header }  from './header/Header';
 export { default as MovieDetails } from './movie-details/MovieDetail';
 export { default as MovieList } from './movie-list/MovieList';
-export { default as MovieElement } from './movie-list/MovieElement';
+export { default as MovieElement } from './movie-list/MovieElement/MovieElement';
 export { default as Loading } from './utils/Loading';
