+++ src/components/movie-details/MovieDetail.js
@@ -4,7 +4,7 @@ export default class MovieDetail extends Component {
 
     render() {
         return (
-            <div className="w-25 border p-4 d-flex flex-column">
+            <div className="w-25 bg-light p-4 d-flex flex-column">
                 <h5>{ this.props.movie.title }</h5>
                 <hr className="w-100"/>
                 <div>
