+++ src/components/movie-list/MovieElement/MovieElement.module.scss
@@ -1,0 +1,11 @@
+.container {
+    width: 300px;
+    height: 200px; 
+    overflow: hidden;
+    margin: 0px 5px 10px 5px;
+    cursor: pointer;
+    &:hover {
+       transform: scale(1.05);
+       transition: all 0.3s;
+    }
+}
