+++ src/components/movie-list/MovieList.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import MovieElement from './MovieElement';
+import MovieElement from './MovieElement/MovieElement';
 
 export default class MovieList extends Component {
 
@@ -7,7 +7,7 @@ export default class MovieList extends Component {
         return (
             <div className="w-75 d-flex flex-row flex-wrap align-content-start">
             { this.props.movies.map( (m, index) => (
-                <MovieElement key={ m.title } movie={ m } updateSelectedMovie={ () => this.props.updateSelectedMovie(index) }/>
+                <MovieElement key={ m.title + index } movie={ m } updateSelectedMovie={ () => this.props.updateSelectedMovie(index) }/>
             ))}     
             </div>
         );
